Add 3 functions
randomGen, which generates a random integer. getComputerChoice, which generates
a random choice between rock, paper and scissors. playRound, which plays a plays
a round of the game based on a player input. This is the core of the program an is very primitive,
now we need to fine tune the messages and add a prompt without case sensitivity.