Add loop that counts the result of 5 games
This is a very primitive version but I believe the core idea of what we
are building is right. We made the return of the playRound() function more
complex, by adding just not a win/loss/tie message, but by assigning 0 or 1
to a player/computer variable. This is hopefully a more convinient way to
keep track of the result and add it to a count variable that adds the results
of 5 games in the loop, for both the player and the computer, and finds who won
more times.