Show alerts with the results in every round
Apart from loging the results, the program now shows them in alerts. This
is useful because it lets the player see the real time score after every round,
it can see what the computer played to make a decision for the next round.
Apart from that, I also cleaned the code and made it more readable and uniform.