Style the page and script.js improvements
This commit takes the previous, minimally styled version and adds a more compelling style to it.
There are also a few improvements in the script.js, for example, I found that the playRound and the
old getScore function could be merged into a single one, that I named getScore. Also, the ending message
has been improved, showing the score, and switching a div class to center it.